requirements.txt
openai
langchain
langchain-openai
faiss-cpu
python-dotenv
tiktoken